LIBFCREPO-905. Modified implementation to use X-Forwarded header
Using the "X-Forwarded-Host" and "X-Forwarded-Proto" headers removes
the need for the Subjects to be modified in the
ProxiedLinkedDataProvider.parseResponse method. It also simplifies the
implementation in other ways.

The "describedBy" link mapping is now handled by the
ProxiedLinkedDataProvider class, instead of (misusing) the "setProperty"
method on the EndPoint. The ProxiedLinkedDataProvider
also removes the mapping, to prevent a memory leak.

Removed and rearranged some of the code.

https://issues.umd.edu/browse/LIBFCREPO-905