LIBFCREPO-942. Ensure stale resource is removed from cache
Since the LdpathProcessor class is only called when inserting items
into Solr, any instances of the resource in the
LDCachingInfinispanBackend cache are "stale" and should be removed, so
that a new copy of the resource will be retrieved.

Without this change, subsequent edits to a resource within 24 hours of
the first change (the default expiration time of the
DCachingInfinispanBackend cache) will be ignored and not reflected in
the JSON sent to Solr.

https://issues.umd.edu/browse/LIBFCREPO-942