LIBFCREPO-949. Refactored LdpathProcessorContext to override methods
Refactored the "LdpathProcessor" class, placing auth token retrieval
into a separate method, so that it can be overridden
by subclasses (such as a "test" subclass that simply returns a string)

Also changed the "getLinkedDataResourceUrl" method to "protected",
so it can be overridden by a subclass, if needed.

In LdpathProcessorContextTest, enabled the tests to pass successfully by
adding a "test" version of the "LdpathProcessor" class, and rewriting
the tests to remove the Spring dependencies.

Removed the Spring dependences from the pom.xml, which were clashing
with the Spring versions provided by the "umd-fcrepo-webapp" dependency.

https://issues.umd.edu/browse/LIBFCREPO-949