LIBFCREPO-949. Refactored Spring functionality out of unit tests
Refactored unit tests to remove dependency on Spring, instead using
regular Java object initialization.

Removed the "LdpathProcessorTest", because all the test code was
commented out, and the code was for an older version of the
LdpathProcessor (so it did not work, even if uncommented).

Renamed the "LdpathProcessorContextTest" to "LdpathProcessorExchangeTest"
which demonstrates how to test by providing a Camel "Exchange" object
to a provider.

Also created a parallel "LdpathProcessorRouteTest", which tests the
same functionality as "LdpathProcessorContextTest, but demonstrates
how to set up a Camel route.

Added "TestUtils" to hold common test utility functions, and
"TestLdpathProcessor" to stub out the authorization and
linkedDataResourceUrl creation.

https://issues.umd.edu/browse/LIBFCREPO-949